feat: Added approach timing to pyodide
The Walrus-iterator approach is actually comparable or
slower than the other approaches when run in pyodide online.
Also when I just run directly with `python` instead of `kernprof`
I also lose the Walrus-iterator speedups. I'm wondering if the
speedup is actually just an artifact of being faster for line_profiler.